# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CarriageLayoutDTO:
    properties:
      code:
        description: |-
          Carriage code
          example: CAR1
        type: string
      seats:
        description: List of seats with their availability
        items:
          $ref: '#/definitions/dto.SeatAvailabilityDTO'
        type: array
    type: object
...
  dto.CreateScheduleRequest:
    properties:
      arrival_time:
        description: |-
          Arrival time in YYYY-MM-DD HH:MM format
          required: true
          example: 2024-01-15 16:45
        type: string
      available_seats:
        description: |-
          Number of available seats
          required: true
          minimum: 0
          example: 50
        minimum: 0
        type: integer
      departure_time:
        description: |-
          Departure time in YYYY-MM-DD HH:MM format
          required: true
          example: 2024-01-15 08:30
        type: string
      destination_station_id:
        description: |-
          Destination station ID
          required: true
          example: 507f1f77bcf86cd799439012
        type: string
      origin_station_id:
        description: |-
          Origin station ID
          required: true
          example: 507f1f77bcf86cd799439011
        type: string
      price:
        description: |-
          Ticket price
          required: true
          minimum: 0
          example: 150000
        type: number
      train_id:
        description: |-
          Train ID
          required: true
          example: 507f1f77bcf86cd799439013
        type: string
    required:
    - arrival_time
    - available_seats
    - departure_time
    - destination_station_id
    - origin_station_id
    - price
    - train_id
    type: object
...
  dto.Response-any:
    properties:
      code:
        description: |-
          HTTP status code
          example: 200
        type: integer
      data:
        description: Response data
      message:
        description: |-
          Response message
          example: Success
        type: string
    type: object
...
  dto.Response-array_dto_CarriageLayoutDTO:
    properties:
      code:
        description: |-
          HTTP status code
          example: 200
        type: integer
      data:
        description: Response data
        items:
          $ref: '#/definitions/dto.CarriageLayoutDTO'
        type: array
      message:
        description: |-
          Response message
          example: Success
        type: string
    type: object
  dto.Response-array_dto_ScheduleResponse:
    properties:
      code:
        description: |-
          HTTP status code
          example: 200
        type: integer
      data:
        description: Response data
        items:
          $ref: '#/definitions/dto.ScheduleResponse'
        type: array
      message:
        description: |-
          Response message
          example: Success
        type: string
    type: object
...
  dto.Response-dto_ScheduleResponse:
    properties:
      code:
        description: |-
          HTTP status code
          example: 200
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.ScheduleResponse'
        description: Response data
      message:
        description: |-
          Response message
          example: Success
        type: string
    type: object
...
  dto.ScheduleResponse:
    properties:
      arrival_time:
        description: |-
          Arrival time in YYYY-MM-DD HH:MM format
          example: 2024-01-15 16:45
        type: string
      available_seats:
        description: |-
          Number of available seats
          example: 25
        type: integer
      departure_time:
        description: |-
          Departure time in YYYY-MM-DD HH:MM format
          example: 2024-01-15 08:30
        type: string
      destination_station:
        allOf:
        - $ref: '#/definitions/dto.StationResponse'
        description: Destination station information
      duration:
        description: |-
          Journey duration (e.g., "8h 15m")
          example: 8h 15m
        type: string
      id:
        description: |-
          Schedule unique identifier
          example: 507f1f77bcf86cd799439014
        type: string
      origin_station:
        allOf:
        - $ref: '#/definitions/dto.StationResponse'
        description: Origin station information
      price:
        description: |-
          Ticket price
          example: 150000
        type: number
      train:
        allOf:
        - $ref: '#/definitions/dto.TrainInScheduleResponse'
        description: Train information
    type: object
  dto.SearchScheduleRequest:
    properties:
      departure_date:
        description: |-
          Departure date in YYYY-MM-DD format
          required: true
          example: 2024-01-15
        type: string
      destination_station_id:
        description: |-
          Destination station ID
          required: true
          example: 507f1f77bcf86cd799439012
        type: string
      origin_station_id:
        description: |-
          Origin station ID
          required: true
          example: 507f1f77bcf86cd799439011
        type: string
      passengers:
        description: |-
          Number of passengers
          required: true
          minimum: 1
          example: 2
        minimum: 1
        type: integer
    required:
    - departure_date
    - destination_station_id
    - origin_station_id
    - passengers
    type: object
  dto.SeatAvailabilityDTO:
    properties:
      code:
        description: |-
          Seat code (e.g., A1, B2)
          example: A1
        type: string
      is_available:
        description: |-
          Whether the seat is available for booking
          example: true
        type: boolean
    type: object
...
  dto.StationResponse:
    properties:
      city:
        description: |-
          City where the station is located
          example: Jakarta
        type: string
      code:
        description: |-
          Station code
          example: GMB
        type: string
      id:
        description: |-
          Station unique identifier
          example: 507f1f77bcf86cd799439011
        type: string
      name:
        description: |-
          Station name
          example: Gambir Station
        type: string
    type: object
  dto.TrainInScheduleResponse:
    properties:
      class:
        description: |-
          Train class
          example: Executive
        type: string
      name:
        description: |-
          Train name
          example: Argo Bromo Anggrek
        type: string
    type: object
...
  dto.UpdateScheduleRequest:
    properties:
      arrival_time:
        description: |-
          Arrival time in YYYY-MM-DD HH:MM format
          required: true
          example: 2024-01-15 16:45
        type: string
      available_seats:
        description: |-
          Number of available seats
          required: true
          minimum: 0
          example: 50
        minimum: 0
        type: integer
      departure_time:
        description: |-
          Departure time in YYYY-MM-DD HH:MM format
          required: true
          example: 2024-01-15 08:30
        type: string
      price:
        description: |-
          Ticket price
          required: true
          minimum: 0
          example: 150000
        type: number
    required:
    - arrival_time
    - available_seats
    - departure_time
    - price
    type: object
...
host: localhost:8080
info:
  contact:
    email: fernandolouis55@gmail.com
    name: Louis Fernando
  description: |-
    This is a comprehensive REST API for Indonesian Railway (KAI) ticket booking system.
    The API provides endpoints for user authentication, station management, train management, schedule search, and booking operations.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  title: KAI Backend
  version: "1.0"
paths:
...
  /schedules:
    get:
      consumes:
      - application/json
      description: Retrieve all train schedules for admin management
      produces:
      - application/json
      responses:
        "200":
          description: All schedules retrieved successfully
          schema:
            $ref: '#/definitions/dto.Response-array_dto_ScheduleResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      security:
      - Bearer: []
      summary: Get all schedules (Admin only)
      tags:
      - Schedules
    post:
      consumes:
      - application/json
      description: Create a new train schedule (Admin only)
      parameters:
      - description: Schedule data
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/dto.CreateScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Schedule created successfully
          schema:
            $ref: '#/definitions/dto.Response-dto_ScheduleResponse'
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/dto.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      security:
      - Bearer: []
      summary: Create a new schedule
      tags:
      - Schedules
  /schedules/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a train schedule (Admin only)
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schedule deleted successfully
          schema:
            $ref: '#/definitions/dto.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      security:
      - Bearer: []
      summary: Delete schedule
      tags:
      - Schedules
    get:
      consumes:
      - application/json
      description: Retrieve a specific train schedule by its ID
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schedule retrieved successfully
          schema:
            $ref: '#/definitions/dto.Response-dto_ScheduleResponse'
        "404":
          description: Schedule not found
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      summary: Get schedule by ID
      tags:
      - Schedules
    put:
      consumes:
      - application/json
      description: Update an existing train schedule (Admin only)
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated schedule data
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Schedule updated successfully
          schema:
            $ref: '#/definitions/dto.Response-dto_ScheduleResponse'
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/dto.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      security:
      - Bearer: []
      summary: Update schedule
      tags:
      - Schedules
  /schedules/{id}/seats:
    get:
      consumes:
      - application/json
      description: Retrieve the seat layout for a specific schedule
      parameters:
      - description: Schedule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Seat layout retrieved successfully
          schema:
            $ref: '#/definitions/dto.Response-array_dto_CarriageLayoutDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      summary: Get seat layout
      tags:
      - Schedules
  /schedules/search:
    post:
      consumes:
      - application/json
      description: Search train schedules by origin, destination, date, and passenger
        count
      parameters:
      - description: Schedule data
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/dto.SearchScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Schedules retrieved successfully
          schema:
            $ref: '#/definitions/dto.Response-array_dto_ScheduleResponse'
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/dto.Response-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.Response-any'
      summary: Search schedules
      tags:
      - Schedules
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: 'JWT Authorization header using the Bearer scheme. Example: "Authorization:
      Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"